requirements.txt
matplotlib>=3.7
numpy>=1.24
requests>=2.28
vaderSentiment>=3.3
yfinance>=0.2
python-dotenv>=1.0
streamlit>=1.30
pandas>=2.0
